Rewrite wait_for test to IsolatedAsyncioTestCase
